feat(api): allowing some credentials to override others
